[UPDATE] Fix migrations type hints and add seeders

+++ database/migrations/2019_09_12_041644_create_refunds_table.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
@@ -11,20 +13,21 @@ class CreateRefundsTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
-        Schema::create('refunds', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('type');
-            $table->string('description');
-            $table->dateTime('date');
-            $table->float('value');
-            $table->unsignedBigInteger('userId');
-            $table->timestamps();
-            /** Constraints */
-            $table->foreign('userId')->references('id')->on('usuarios')->onDelete('cascade');
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('refunds')) {
+            Schema::create('refunds', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('type');
+                $table->string('description');
+                $table->float('value');
+                $table->unsignedBigInteger('userId');
+                $table->timestamps();
+                /** Constraints */
+                $table->foreign('userId')->references('id')->on('usuarios')->onDelete('cascade');
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -32,7 +35,7 @@ class CreateRefundsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('refunds');
     }
